SurfaceQuasiGeostrophic

Documentation for SurfaceQuasiGeostrophic.

The unknown of the model is the buoyancy $b(t,x,y)$ and the stream function $\psi(t,x,y)$ which are related through a Poisson like equation $| −i \nabla | \psi = b$, which becomes in Fourier $|k| \hat{\psi} = \hat{b}$

The dynamical model is the following

\[\frac{\partial b}{\partial t} + (\nabla \times \psi) \cdot \nabla b = 0. \]

The solver uses Fourier techniques so that

\[\nabla \times \psi \cdot \nabla b = ifft[(ik) \times \hat{\psi}] \cdot ifft[(ik)\hat{b}],\]

SurfaceQuasiGeostrophic.Grid

SurfaceQuasiGeostrophic.Poisson

SurfaceQuasiGeostrophic.SQG

SurfaceQuasiGeostrophic.compute_dt

SurfaceQuasiGeostrophic.compute_hypercoef

SurfaceQuasiGeostrophic.compute_psi!

SurfaceQuasiGeostrophic.compute_velocities!

SurfaceQuasiGeostrophic.init_buoyancy!

SurfaceQuasiGeostrophic.step!

SurfaceQuasiGeostrophic.update_advection_term!

SurfaceQuasiGeostrophic.update_velocities!

SurfaceQuasiGeostrophic.Grid — Type

init_grid_k(model)

Create a grid in the Fourier space

SurfaceQuasiGeostrophic.Poisson — Type

Poisson( grid )

Derived type to solve the Poisson equation on 2d regular cartesian mesh with periodic boundary conditions on both sides

kx : wave number in x

ky : wave number in y

k2 : $k_x^2 + k_y^2$

nc_x : cells number in x

nc_y : cells number in y

dx : x step size

dy : y step size

rht : fft(rho)

SurfaceQuasiGeostrophic.SQG — Type

SQG

Data structure to store arrays and parameters for the Surface-Quasi-Geostrophic model

rho : Background density

g : Gravity

buoyancyfreqn : Background stratification

odg_b : Amplitude of the buoyancy

SurfaceQuasiGeostrophic.compute_dt — Method

compute_dt( model, ogird, w)

CFL of the (large-scale) advection

SurfaceQuasiGeostrophic.compute_hypercoef — Method

compute_hypercoef( model )

compute hyperviscosity coefficient

SurfaceQuasiGeostrophic.compute_psi! — Method

compute_psi!( poisson, psi, omega )

computes $\psi$ from $\omega$

\[-\Delta \psi(x,y) = \omega(x,y)\]

SurfaceQuasiGeostrophic.compute_velocities! — Method

compute_velocities!( ux, uy, poisson, omega )

solves Poisson equation to compute velocity fields

\[\Delta \psi(x,y) = - \omega(x,y) \\ u(x,y) = \nabla \times \psi(x,y) \]

SurfaceQuasiGeostrophic.init_buoyancy! — Method

init_buoyancy!(model)

Initial condition for the buoyancy

SurfaceQuasiGeostrophic.step! — Method

step!(model, time_solver, dt)

Time integration of the advection equation of b by 4th order Runge-Kutta method

SurfaceQuasiGeostrophic.update_advection_term! — Method

update_advection_term(model)

Compute the Fourier transform of the partial derivative of the advection term plus the hyperviscosity term

SurfaceQuasiGeostrophic.update_velocities! — Method

update_velocities!(u_x, u_y, model, b̂)

Compute the streamfunction and the velocity from the buoyancy according to the Surface-Quasi-Geostrophic model.